auth/requirements.txt
defusedxml==0.4.1
Django==1.8.5
django-allauth==0.23.0
django-cors-headers==1.1.0
django-rest-auth==0.5.0
djangorestframework==3.2.4
djangorestframework-jwt==1.7.2
oauthlib==1.0.3
PyJWT==1.4.0
python3-openid==3.0.7
requests==2.8.1
requests-oauthlib==0.5.0
six==1.10.0
wheel==0.24.0
